create wrapper binary to call some commands
'aa-caller' is a new wrapper binary to handle calls to 'aa-status', 'aa-unconfined', and 'dmesg'. This limits the number of pkexec authentication prompts the AppAnvil user will see. Instead of getting a prompt individually for each of these calls, calling 'aa-status' will ensure there is only one prompt when the program starts. This brings the behavior of 'pkexec' closer to 'sudo' for these three commands.