Removed old tests and wrote new ones
I removed the old tests as they were heavily reliant on deque being used so it was moreso a test for the std::deque library than for the blocking_queue class. Now that dependency injection has been completed, we can test blocking_queue by simply assuring it calls the injected deque object since we can assume the standard library is sufficiently tested already. The program compiles again as of the last commit, however the tests for blocking_queue still fail. I believe this is due to the mock deque object's expect_calls are deleted when the queue is passed from BlockingQueueTest -> BlockingQueueMock -> BlockingQueue. I attempted to solve this by passing the queue using std::shared_ptr as is done for the other mocked object in existing tests, but this resulted in a segmentation fault when attempting to run the application, so I figured I should just leave it as it is in this commit until this issue is resolved since this version doesn't break the application.