Added filtering capability for child processes
Updated Status::filter to allow for filtering to recognize child processes in new hierarchical view.
Fixed the error that was thrown when filtering processes, which was actually not introduced by the updates to processes.cc in the last commit. This build is stable and works but could still use some improvements, such as causing parent processes to automatically expand when the search filter matches a child process, and causing child processes of a search match to remain shown even if they do not match the search filter.
Also made a small change in logs.inl to avoid redundancy