improve layout of 'resources/' directory
this change restructures the layout of './resources' to make more sense. Instead of everything being in one folder, I moved the ui files into subdirectories. This then required me to make changes to the build and codebase. Furthermore, all generated files in the resources directory will now be outputted to the 'resources/generated' subdirectory